BAR-1208: Refactored away duplicate code by introducing OppfolgingTypeContract
