+++ src/main/kotlin/no/ssb/barn/xsd/BarnevernType.kt
@@ -1,8 +1,8 @@
 package no.ssb.barn.xsd
 
+import jakarta.xml.bind.annotation.*
 import java.time.ZonedDateTime
 import java.util.UUID
-import jakarta.xml.bind.annotation.*
 
 @XmlRootElement(name = "Barnevern")
 @XmlAccessorType(XmlAccessType.FIELD)
@@ -28,5 +28,5 @@ data class BarnevernType(
     val avgiver: AvgiverType,
 
     @field:XmlElement(name = "Sak", required = true)
-    val sak: SakType
+    val sak: SakType<OppfolgingType>
 )
