+++ src/main/kotlin/no/ssb/barn/xsd/OppfolgingTypeContract.kt
@@ -1,0 +1,8 @@
+package no.ssb.barn.xsd
+
+/**
+ * Empty OppfolgingTypeContract interface for narrowing down the type
+ * when [OppfolgingType] and [OppfolgingTypeV3] are used as generic.
+ */
+interface OppfolgingTypeContract {
+}
